cleanup round.c functions from game.c
