+++ source/game.c
@@ -46,13 +46,6 @@
 
 #define EXPIRE_ANIMATION_FRAME_COUNT 3
 
-typedef struct
-{
-    u32 chips;
-    u32 mult;
-    char* display_name;
-} HandValues;
-
 // Used as a No Operation for game states that have no init and/or exit function.
 // ricfehr3 did the work of determining whether a noop or a NULL check was more
 // efficient. Well, this is the answer.
@@ -70,95 +63,6 @@ StateInfo state_info[] = {
     {.on_init = init_fn, .on_update = update_fn, .on_exit = exit_fn, .substate = 0},
 #include "../include/def_state_info_table.h"
 #undef DEF_STATE_INFO
-};
-
-extern int jokers_sel_row_get_size(void);
-extern bool jokers_sel_row_on_selection_changed(
-    SelectionGrid* selection_grid,
-    int row_idx,
-    const Selection* prev_selection,
-    const Selection* new_selection
-);
-extern void jokers_sel_row_on_key_transit(SelectionGrid* selection_grid, Selection* selection);
-
-static int game_playing_hand_row_get_size(void);
-extern bool game_playing_hand_row_on_selection_changed(
-    SelectionGrid* selection_grid,
-    int row_idx,
-    const Selection* prev_selection,
-    const Selection* new_selection
-);
-extern void game_playing_hand_row_on_key_transit(
-    SelectionGrid* selection_grid,
-    Selection* selection
-);
-
-int game_playing_button_row_get_size(void);
-extern bool game_playing_button_row_on_selection_changed(
-    SelectionGrid* selection_grid,
-    int row_idx,
-    const Selection* prev_selection,
-    const Selection* new_selection
-);
-extern void game_playing_button_row_on_key_hit(SelectionGrid* selection_grid, Selection* selection);
-
-// clang-format off
-SelectionGridRow game_playing_selection_rows[] = {
-    {
-        0,
-        jokers_sel_row_get_size,
-        jokers_sel_row_on_selection_changed,
-        jokers_sel_row_on_key_transit,
-        {.wrap = false}
-    },
-    {
-        1,
-        game_playing_hand_row_get_size,
-        game_playing_hand_row_on_selection_changed,
-        game_playing_hand_row_on_key_transit,
-        {.wrap = true}
-    },
-    {
-        2,
-        game_playing_button_row_get_size,
-        game_playing_button_row_on_selection_changed,
-        game_playing_button_row_on_key_hit,
-        {.wrap = false}
-    }
-};
-// clang-format on
-
-SelectionGrid game_playing_selection_grid = {
-    game_playing_selection_rows,
-    NUM_ELEM_IN_ARR(game_playing_selection_rows),
-    GAME_PLAYING_INIT_SEL
-};
-
-static void game_playing_play_hand_on_pressed(void);
-static void game_playing_discard_on_pressed(void);
-extern bool can_play_hand(void);
-static bool can_discard_hand(void);
-// Array of buttons by horizontal selection index (x)
-Button game_playing_buttons[] = {
-    {PLAY_HAND_BTN_BORDER_PID, PLAY_HAND_BTN_PID, game_playing_play_hand_on_pressed, can_play_hand   },
-    {DISCARD_BTN_BORDER_PID,   DISCARD_BTN_PID,   game_playing_discard_on_pressed,   can_discard_hand},
-};
-
-static const HandValues hand_base_values[] = {
-    {.chips = 0,   .mult = 0,  .display_name = NULL     }, // NONE
-    {.chips = 5,   .mult = 1,  .display_name = "HIGH C" }, // HIGH_CARD
-    {.chips = 10,  .mult = 2,  .display_name = "PAIR"   }, // PAIR
-    {.chips = 20,  .mult = 2,  .display_name = "2 PAIR" }, // TWO_PAIR
-    {.chips = 30,  .mult = 3,  .display_name = "3 OAK"  }, // THREE_OF_A_KIND
-    {.chips = 60,  .mult = 7,  .display_name = "4 OAK"  }, // FOUR_OF_A_KIND
-    {.chips = 30,  .mult = 4,  .display_name = "STRT"   }, // STRAIGHT
-    {.chips = 35,  .mult = 4,  .display_name = "FLUSH"  }, // FLUSH
-    {.chips = 40,  .mult = 4,  .display_name = "FULL H" }, // FULL_HOUSE
-    {.chips = 100, .mult = 8,  .display_name = "STRT F" }, // STRAIGHT_FLUSH
-    {.chips = 100, .mult = 8,  .display_name = "ROYAL F"}, // ROYAL_FLUSH
-    {.chips = 120, .mult = 12, .display_name = "5 OAK"  }, // FIVE_OF_A_KIND
-    {.chips = 140, .mult = 14, .display_name = "FLUSH H"}, // FLUSH_HOUSE
-    {.chips = 160, .mult = 16, .display_name = "FLUSH 5"}  // FLUSH_FIVE
 };
 
 // The current game state, this is used to determine what the game is doing at any given time
@@ -209,7 +113,6 @@ bool retrigger = false;
 
 int hand_size = 8; // Default hand size is 8
 int cards_drawn = 0;
-int hand_selections = 0;
 
 // Keeping track of cards scored
 int scored_card_index = 0;
@@ -256,50 +159,6 @@ Bitset* get_avail_jokers_bitset_ptr(void)
     return &_avail_jokers_bitset;
 }
 
-// Round.c
-
-void played_push(CardObject* card_object)
-{
-    if (played_top >= MAX_SELECTION_SIZE - 1)
-        return;
-    played[++played_top] = card_object;
-}
-
-static inline CardObject* played_pop()
-{
-    if (played_top < 0)
-        return NULL;
-    return played[played_top--];
-}
-
-void deck_push(Card* card)
-{
-    if (deck_top >= MAX_DECK_SIZE - 1)
-        return;
-    deck[++deck_top] = card;
-}
-
-Card* deck_pop()
-{
-    if (deck_top < 0)
-        return NULL;
-    return deck[deck_top--];
-}
-
-void discard_push(Card* card)
-{
-    if (discard_top >= MAX_DECK_SIZE - 1)
-        return;
-    discard_pile[++discard_top] = card;
-}
-
-Card* discard_pop()
-{
-    if (discard_top < 0)
-        return NULL;
-    return discard_pile[discard_top--];
-}
-
 // ============================================================================
 // Getter and Setter Functions
 // ============================================================================
@@ -308,6 +167,24 @@ Card* discard_pop()
 CardObject** get_hand_array(void)
 {
     return hand;
+}
+
+CardObject* get_hand_card_at(int idx)
+{
+    if (idx < 0 || idx > hand_top)
+    {
+        return NULL;
+    }
+    return hand[idx];
+}
+
+void set_hand_card_at(int idx, CardObject* card_object)
+{
+    if (idx < 0 || idx > hand_top)
+    {
+        return;
+    }
+    hand[idx] = card_object;
 }
 
 int get_hand_top(void)
@@ -330,6 +207,11 @@ int get_played_top(void)
     return played_top;
 }
 
+int increment_played_top(void)
+{
+    return ++played_top;
+}
+
 int get_scored_card_index(void)
 {
     return scored_card_index;
@@ -338,6 +220,18 @@ int get_scored_card_index(void)
 int get_deck_top(void)
 {
     return deck_top;
+}
+
+int increment_deck_top(void)
+{
+    return ++deck_top;
+}
+
+void set_deck_at(int idx, Card* card)
+{
+    if (idx < 0 || idx > MAX_DECK_SIZE - 1)
+        return;
+    deck[idx] = card;
 }
 
 int deck_get_size(void)
@@ -349,6 +243,20 @@ int deck_get_max_size(void)
 {
     // This is the max amount of cards that the player currently has in their possession
     return hand_top + played_top + deck_top + discard_top + 4;
+}
+
+void set_played_card_at(int idx, CardObject* card_object)
+{
+    if (idx < 0 || idx >= MAX_SELECTION_SIZE)
+        return;
+    played[idx] = card_object;
+}
+
+CardObject* get_played_card_at(int idx)
+{
+    if (idx < 0 || idx >= MAX_SELECTION_SIZE)
+        return NULL;
+    return played[idx];
 }
 
 // Joker List Getters
@@ -478,6 +386,11 @@ void reset_hands(void)
     hands = max_hands;
 }
 
+int decrement_hands(void)
+{
+    return --hands;
+}
+
 int get_num_discards_remaining(void)
 {
     return discards;
@@ -491,6 +404,11 @@ int get_discards(void)
 void reset_discards(void)
 {
     discards = max_discards;
+}
+
+int decrement_discards(void)
+{
+    return --discards;
 }
 
 // Round and Ante
@@ -670,8 +588,63 @@ void mult_rng_seed(int factor)
     rng_seed *= factor;
 }
 
-// End of Getter and Setter Functions
+// Hand state
+
+void set_hand_state(enum HandState new_hand_state)
+{
+    hand_state = new_hand_state;
+}
+
+enum HandState get_hand_state(void)
+{
+    return hand_state;
+}
+
 // ============================================================================
+// Round.c helpers
+// ============================================================================
+
+void played_push(CardObject* card_object)
+{
+    if (played_top >= MAX_SELECTION_SIZE - 1)
+        return;
+    played[++played_top] = card_object;
+}
+
+CardObject* played_pop()
+{
+    if (played_top < 0)
+        return NULL;
+    return played[played_top--];
+}
+
+void deck_push(Card* card)
+{
+    if (deck_top >= MAX_DECK_SIZE - 1)
+        return;
+    deck[++deck_top] = card;
+}
+
+Card* deck_pop()
+{
+    if (deck_top < 0)
+        return NULL;
+    return deck[deck_top--];
+}
+
+void discard_push(Card* card)
+{
+    if (discard_top >= MAX_DECK_SIZE - 1)
+        return;
+    discard_pile[++discard_top] = card;
+}
+
+Card* discard_pop()
+{
+    if (discard_top < 0)
+        return NULL;
+    return discard_pile[discard_top--];
+}
 
 // ============================================================================
 // Joker Query Functions
@@ -740,227 +713,6 @@ void increment_four_fingers_joker_count(void)
 void increment_shortcut_joker_count(void)
 {
     shortcut_joker_count++;
-}
-
-// ============================================================================
-// Hand Management Functions
-// ============================================================================
-
-// idx_a and idx_b are assumed to be valid indexes within the hand array
-// no checks will be performed here for performance's sake
-void swap_cards_in_hand(int idx_a, int idx_b)
-{
-    CardObject* temp = hand[idx_a];
-    hand[idx_a] = hand[idx_b];
-    hand[idx_b] = temp;
-}
-
-void sort_hand_by_suit(void)
-{
-    for (int idx_a = 0; idx_a < hand_top; idx_a++)
-    {
-        for (int idx_b = idx_a + 1; idx_b <= hand_top; idx_b++)
-        {
-            if (hand[idx_a] == NULL ||
-                (hand[idx_b] != NULL && (hand[idx_a]->card->suit > hand[idx_b]->card->suit ||
-                                         (hand[idx_a]->card->suit == hand[idx_b]->card->suit &&
-                                          hand[idx_a]->card->rank > hand[idx_b]->card->rank))))
-            {
-                swap_cards_in_hand(idx_a, idx_b);
-            }
-        }
-    }
-}
-
-void sort_hand_by_rank(void)
-{
-    for (int idx_a = 0; idx_a < hand_top; idx_a++)
-    {
-        for (int idx_b = idx_a + 1; idx_b <= hand_top; idx_b++)
-        {
-            if (hand[idx_a] == NULL ||
-                (hand[idx_b] != NULL && hand[idx_a]->card->rank > hand[idx_b]->card->rank))
-            {
-                swap_cards_in_hand(idx_a, idx_b);
-            }
-        }
-    }
-}
-
-bool shift_null_card_to_end(int null_card_idx)
-{
-    // Start by searching any non NULL cards after the NULL one
-    // don't start at null_card_idx+1 to avoid potential illegal array access
-    int non_null_card_idx = null_card_idx;
-    for (; non_null_card_idx <= hand_top; non_null_card_idx++)
-    {
-        if (hand[non_null_card_idx] != NULL)
-        {
-            break;
-        }
-    }
-
-    // return false if there are no non-NULL cards left/there are no more sprites to destroy
-    if (non_null_card_idx > hand_top)
-    {
-        return false;
-    }
-
-    // If there is one, shift it and all the cards that follow forward
-    // This way we close the gap and ensure the next card is not NULL
-
-    // Iterating up to `hand_top - non_null_card_idx + 1` should end up out of bounds
-    // but for some reason it doesn't pose any issue, and taking out the +1 breaks
-    // the code, so I'll be elaving it here until someone figures it out ^^'
-    for (int j = 0; j <= hand_top - non_null_card_idx + 1; j++)
-    {
-        hand[null_card_idx + j] = hand[non_null_card_idx + j];
-    }
-
-    return true;
-}
-
-// ============================================================================
-// Hand Type Detection and Card Evaluation
-// ============================================================================
-
-static inline enum HandType hand_get_type(void)
-{
-    enum HandType res_hand_type = NONE;
-
-    // Idk if this is how Balatro does it but this is how I'm doing it
-    if (hand_selections == 0 || hand_state == HAND_DISCARD)
-    {
-        res_hand_type = NONE;
-        return res_hand_type;
-    }
-
-    res_hand_type = HIGH_CARD;
-
-    u8 suits[NUM_SUITS];
-    u8 ranks[NUM_RANKS];
-    get_hand_distribution(ranks, suits);
-
-    // Check for flush
-    if (hand_contains_flush(suits))
-        res_hand_type = FLUSH;
-
-    // Check for straight
-    if (hand_contains_straight(ranks))
-    {
-        if (res_hand_type == FLUSH)
-            res_hand_type = STRAIGHT_FLUSH;
-        else
-            res_hand_type = STRAIGHT;
-    }
-
-    // The following can be optimized better but not sure how much it matters
-    u8 n_of_a_kind = hand_contains_n_of_a_kind(ranks);
-
-    if (n_of_a_kind >= 5)
-    {
-        if (res_hand_type == FLUSH)
-        {
-            return FLUSH_FIVE;
-        }
-        return FIVE_OF_A_KIND;
-    }
-
-    // Check for royal flush vs regular straight flush
-    if (res_hand_type == STRAIGHT_FLUSH)
-    {
-        if (ranks[TEN] && ranks[JACK] && ranks[QUEEN] && ranks[KING] && ranks[ACE])
-            return ROYAL_FLUSH;
-        return STRAIGHT_FLUSH;
-    }
-
-    if (n_of_a_kind == 4)
-    {
-        return FOUR_OF_A_KIND;
-    }
-
-    if (n_of_a_kind == 3 && hand_contains_full_house(ranks))
-    {
-        return FULL_HOUSE;
-    }
-
-    // Flush and Straight are more valuable than the remaining hand types, so return them now
-    if (res_hand_type == FLUSH)
-    {
-        if (n_of_a_kind >= 5)
-        {
-            return FLUSH_HOUSE;
-        }
-        return FLUSH;
-    }
-    if (res_hand_type == STRAIGHT)
-    {
-        return STRAIGHT;
-    }
-
-    if (n_of_a_kind == 3)
-    {
-        return THREE_OF_A_KIND;
-    }
-
-    if (n_of_a_kind == 2)
-    {
-        if (hand_contains_two_pair(ranks))
-        {
-            return TWO_PAIR;
-        }
-        return PAIR;
-    }
-
-    return res_hand_type; // should be HIGH_CARD
-}
-
-static void print_hand_type(const char* hand_type_str)
-{
-    if (hand_type_str == NULL)
-        return; // NULL-checking paranoia
-    tte_printf(
-        "#{P:%d,%d; cx:0x%X000}%s",
-        HAND_TYPE_RECT.left,
-        HAND_TYPE_RECT.top,
-        TTE_WHITE_PB,
-        hand_type_str
-    );
-}
-
-void set_hand(void)
-{
-    tte_erase_rect_wrapper(HAND_TYPE_RECT);
-    hand_type = hand_get_type();
-
-    HandValues hand = hand_base_values[hand_type];
-
-    chips = hand.chips;
-    mult = hand.mult;
-
-    print_hand_type(hand.display_name);
-    display_chips();
-    display_mult();
-}
-
-static bool can_discard_hand(void)
-{
-    return (discards > 0 && hand_state == HAND_SELECT && hand_selections > 0);
-}
-
-// ============================================================================
-// Deck Management Functions
-// ============================================================================
-
-void deck_shuffle(void)
-{
-    for (int i = deck_top; i > 0; i--)
-    {
-        int j = rand() % (i + 1);
-        Card* temp = deck[i];
-        deck[i] = deck[j];
-        deck[j] = temp;
-    }
 }
 
 // ============================================================================
@@ -1135,51 +887,6 @@ static inline void set_seed(int seed)
 {
     rng_seed = seed;
     srand(rng_seed);
-}
-
-// Playing state functions
-static void game_playing_discard_on_pressed(void)
-{
-    if (!can_discard_hand())
-        return;
-
-    hand_state = HAND_DISCARD;
-    discards--;
-    display_hands();
-    set_hand();
-    tte_printf(
-        "#{P:%d,%d; cx:0x%X000}%d",
-        DISCARDS_TEXT_RECT.left,
-        DISCARDS_TEXT_RECT.top,
-        TTE_RED_PB,
-        discards
-    );
-
-    // Move back to hand selection
-    selection_grid_move_selection_vert(&game_playing_selection_grid, -1);
-}
-
-static void game_playing_play_hand_on_pressed(void)
-{
-    if (!can_play_hand())
-        return;
-
-    hand_state = HAND_PLAY;
-    hands--;
-    display_hands();
-
-    // Move back to hand selection
-    selection_grid_move_selection_vert(&game_playing_selection_grid, -1);
-}
-
-static int game_playing_hand_row_get_size(void)
-{
-    return hand_get_size();
-}
-
-int game_playing_button_row_get_size(void)
-{
-    return NUM_ELEM_IN_ARR(game_playing_buttons);
 }
 
 void game_start(void)
