+++ source/game/round.c
@@ -12,6 +12,7 @@
 #include "game/point.h"
 #include "game/rect.h"
 #include "game/selection.h"
+#include "game/shop.h"
 #include "game/timer.h"
 #include "graphic_utils.h"
 #include "hand_analysis.h"
@@ -26,8 +27,6 @@
 #include <stdlib.h>
 #include <tonc.h>
 #include <tonc_memdef.h>
-
-#define HIGHLIGHT_COLOR 0xFFFF
 
 // Sound pitch steps for SFX
 #define PITCH_STEP_DISCARD_SFX   (-64)
@@ -58,72 +57,33 @@
 static const int HAND_SPACING_LUT[MAX_HAND_SIZE] =
     {28, 28, 28, 28, 27, 21, 18, 15, 13, 12, 10, 9, 9, 8, 8, 7};
 
-// Forward declarations - extern functions and constants from game.c
-extern void game_change_state(enum GameState new_game_state);
-extern int game_playing_button_row_get_size(void);
-extern void change_background(enum BackgroundId id);
-extern void deck_shuffle(void);
-extern void sort_hand_by_suit(void);
-extern void sort_hand_by_rank(void);
-extern void set_hand(void);
-extern void display_score(u32 value);
-extern void display_temp_score(u32 value);
-extern void display_chips(void);
-extern void display_mult(void);
-extern void display_round(int value);
-extern bool shift_null_card_to_end(int idx);
-extern Card* deck_pop(void);
-extern void deck_push(Card* card);
-extern void discard_push(Card* card);
-extern Card* discard_pop(void);
-extern int deck_get_size(void);
-extern int deck_get_max_size(void);
-extern int hand_get_size(void);
-extern int get_hand_top(void);
-extern int get_straight_and_flush_size(void);
-extern bool is_shortcut_joker_active(void);
-extern void swap_cards_in_hand(int idx1, int idx2);
-extern void increment_blind(enum BlindState increment_reason);
-extern void played_push(CardObject* card);
-extern int find_flush_in_played_cards(
-    CardObject** played,
-    int top,
-    int min_len,
-    bool* out_selection
-);
-extern int find_straight_in_played_cards(
-    CardObject** played,
-    int top,
-    bool shortcut,
-    int min_len,
-    bool* out_selection
-);
-extern void select_paired_cards_in_hand(CardObject* played[], int played_top, bool out[]);
-extern void main_bg_se_copy_rect(Rect src_rect, BG_POINT dest_pos);
-extern void main_bg_se_copy_expand_3x3_rect(Rect dest_rect, BG_POINT src_pos);
-extern void main_bg_se_clear_rect(Rect rect);
-extern void tte_erase_rect_wrapper(Rect rect);
-extern void tte_erase_screen(void);
-extern void play_sfx(mm_word id, mm_word rate, mm_byte volume);
-extern u32 u32_protected_add(u32 a, u32 b);
-extern u32 u32_protected_mult(u32 a, u32 b);
-extern void card_object_shake(CardObject* card_object, mm_word sound_id);
-extern void button_press(Button* button);
+static int hand_selections = 0;
 
-extern const Rect BLIND_REQ_TEXT_RECT;
-extern const Rect DISCARDS_TEXT_RECT;
-extern const Rect DECK_SIZE_RECT;
-extern const Rect ANTE_TEXT_RECT;
-extern const Rect TEMP_SCORE_RECT;
-extern const Rect HAND_SIZE_RECT_SELECT;
-extern const Rect HAND_SIZE_RECT_PLAYING;
-extern const Rect SCORE_FLAME_FRAMES_START;
-extern const BG_POINT SCORE_FLAME_CHIPS_POS;
-extern const BG_POINT SCORE_FLAME_MULT_POS;
+typedef struct
+{
+    u32 chips;
+    u32 mult;
+    char* display_name;
+} HandValues;
+
+static const HandValues hand_base_values[] = {
+    {.chips = 0,   .mult = 0,  .display_name = NULL     }, // NONE
+    {.chips = 5,   .mult = 1,  .display_name = "HIGH C" }, // HIGH_CARD
+    {.chips = 10,  .mult = 2,  .display_name = "PAIR"   }, // PAIR
+    {.chips = 20,  .mult = 2,  .display_name = "2 PAIR" }, // TWO_PAIR
+    {.chips = 30,  .mult = 3,  .display_name = "3 OAK"  }, // THREE_OF_A_KIND
+    {.chips = 60,  .mult = 7,  .display_name = "4 OAK"  }, // FOUR_OF_A_KIND
+    {.chips = 30,  .mult = 4,  .display_name = "STRT"   }, // STRAIGHT
+    {.chips = 35,  .mult = 4,  .display_name = "FLUSH"  }, // FLUSH
+    {.chips = 40,  .mult = 4,  .display_name = "FULL H" }, // FULL_HOUSE
+    {.chips = 100, .mult = 8,  .display_name = "STRT F" }, // STRAIGHT_FLUSH
+    {.chips = 100, .mult = 8,  .display_name = "ROYAL F"}, // ROYAL_FLUSH
+    {.chips = 120, .mult = 12, .display_name = "5 OAK"  }, // FIVE_OF_A_KIND
+    {.chips = 140, .mult = 14, .display_name = "FLUSH H"}, // FLUSH_HOUSE
+    {.chips = 160, .mult = 16, .display_name = "FLUSH 5"}  // FLUSH_FIVE
+};
 
 // Global variables from game.c
-extern enum BackgroundId background;
-extern uint timer;
 extern CardObject* hand[MAX_HAND_SIZE];
 extern int hand_top;
 extern CardObject* played[MAX_SELECTION_SIZE];
@@ -145,7 +105,6 @@ extern FIXED lerped_temp_score;
 extern bool retrigger;
 extern int hand_size;
 extern int cards_drawn;
-extern int hand_selections;
 extern int scored_card_index;
 extern bool sound_played;
 extern bool discarded_card;
@@ -167,10 +126,323 @@ extern ListItr _joker_card_scored_end_itr;
 extern ListItr _joker_round_end_itr;
 
 // Forward declarations for static functions defined later in this file
+static void set_hand(void);
 static void hand_deselect_all_cards(void);
 static void hand_change_sort(void);
 static void hand_select_card(int index);
 static inline int hand_sel_idx_to_card_idx(int selection_index);
+
+static void game_playing_play_hand_on_pressed(void);
+static void game_playing_discard_on_pressed(void);
+static bool can_play_hand(void);
+static bool can_discard_hand(void);
+// Array of buttons by horizontal selection index (x)
+Button game_playing_buttons[] = {
+    {PLAY_HAND_BTN_BORDER_PID, PLAY_HAND_BTN_PID, game_playing_play_hand_on_pressed, can_play_hand   },
+    {DISCARD_BTN_BORDER_PID,   DISCARD_BTN_PID,   game_playing_discard_on_pressed,   can_discard_hand},
+};
+
+static int game_playing_hand_row_get_size(void);
+static bool game_playing_hand_row_on_selection_changed(
+    SelectionGrid* selection_grid,
+    int row_idx,
+    const Selection* prev_selection,
+    const Selection* new_selection
+);
+static void game_playing_hand_row_on_key_transit(
+    SelectionGrid* selection_grid,
+    Selection* selection
+);
+
+static int game_playing_button_row_get_size(void);
+static bool game_playing_button_row_on_selection_changed(
+    SelectionGrid* selection_grid,
+    int row_idx,
+    const Selection* prev_selection,
+    const Selection* new_selection
+);
+static void game_playing_button_row_on_key_hit(SelectionGrid* selection_grid, Selection* selection);
+
+// clang-format off
+SelectionGridRow game_playing_selection_rows[] = {
+    {
+        0,
+        jokers_sel_row_get_size,
+        jokers_sel_row_on_selection_changed,
+        jokers_sel_row_on_key_transit,
+        {.wrap = false}
+    },
+    {
+        1,
+        game_playing_hand_row_get_size,
+        game_playing_hand_row_on_selection_changed,
+        game_playing_hand_row_on_key_transit,
+        {.wrap = true}
+    },
+    {
+        2,
+        game_playing_button_row_get_size,
+        game_playing_button_row_on_selection_changed,
+        game_playing_button_row_on_key_hit,
+        {.wrap = false}
+    }
+};
+// clang-format on
+
+SelectionGrid game_playing_selection_grid = {
+    game_playing_selection_rows,
+    NUM_ELEM_IN_ARR(game_playing_selection_rows),
+    GAME_PLAYING_INIT_SEL
+};
+
+static int game_playing_hand_row_get_size(void)
+{
+    return hand_get_size();
+}
+
+static void game_playing_play_hand_on_pressed(void)
+{
+    if (!can_play_hand())
+        return;
+
+    set_hand_state(HAND_PLAY);
+    decrement_hands();
+    display_hands();
+
+    // Move back to hand selection
+    selection_grid_move_selection_vert(&game_playing_selection_grid, -1);
+}
+
+// Playing state functions
+static void game_playing_discard_on_pressed(void)
+{
+    if (!can_discard_hand())
+        return;
+
+    set_hand_state(HAND_DISCARD);
+    decrement_discards();
+    display_hands();
+    set_hand();
+    tte_printf(
+        "#{P:%d,%d; cx:0x%X000}%d",
+        DISCARDS_TEXT_RECT.left,
+        DISCARDS_TEXT_RECT.top,
+        TTE_RED_PB,
+        discards
+    );
+
+    // Move back to hand selection
+    selection_grid_move_selection_vert(&game_playing_selection_grid, -1);
+}
+
+static int game_playing_button_row_get_size(void)
+{
+    return NUM_ELEM_IN_ARR(game_playing_buttons);
+}
+
+// idx_a and idx_b are assumed to be valid indexes within the hand array
+// no checks will be performed here for performance's sake
+static void swap_cards_in_hand(int idx_a, int idx_b)
+{
+    CardObject* temp = get_hand_card_at(idx_a);
+    set_hand_card_at(idx_a, get_hand_card_at(idx_b));
+    set_hand_card_at(idx_b, temp);
+}
+
+static bool shift_null_card_to_end(int null_card_idx)
+{
+    // Start by searching any non NULL cards after the NULL one
+    // don't start at null_card_idx+1 to avoid potential illegal array access
+    int non_null_card_idx = null_card_idx;
+    for (; non_null_card_idx <= hand_top; non_null_card_idx++)
+    {
+        if (get_hand_card_at(non_null_card_idx) != NULL)
+        {
+            break;
+        }
+    }
+
+    // return false if there are no non-NULL cards left/there are no more sprites to destroy
+    if (non_null_card_idx > hand_top)
+    {
+        return false;
+    }
+
+    // If there is one, shift it and all the cards that follow forward
+    // This way we close the gap and ensure the next card is not NULL
+
+    // Iterating up to `hand_top - non_null_card_idx + 1` should end up out of bounds
+    // but for some reason it doesn't pose any issue, and taking out the +1 breaks
+    // the code, so I'll be elaving it here until someone figures it out ^^'
+    for (int j = 0; j <= hand_top - non_null_card_idx + 1; j++)
+    {
+        set_hand_card_at(null_card_idx + j, get_hand_card_at(non_null_card_idx + j));
+    }
+
+    return true;
+}
+
+void deck_shuffle(void)
+{
+    for (int i = deck_top; i > 0; i--)
+    {
+        int j = rand() % (i + 1);
+        Card* temp = deck[i];
+        deck[i] = deck[j];
+        deck[j] = temp;
+    }
+}
+
+void sort_hand_by_suit(void)
+{
+    for (int idx_a = 0; idx_a < hand_top; idx_a++)
+    {
+        for (int idx_b = idx_a + 1; idx_b <= hand_top; idx_b++)
+        {
+            if (hand[idx_a] == NULL ||
+                (hand[idx_b] != NULL && (hand[idx_a]->card->suit > hand[idx_b]->card->suit ||
+                                         (hand[idx_a]->card->suit == hand[idx_b]->card->suit &&
+                                          hand[idx_a]->card->rank > hand[idx_b]->card->rank))))
+            {
+                swap_cards_in_hand(idx_a, idx_b);
+            }
+        }
+    }
+}
+
+void sort_hand_by_rank(void)
+{
+    for (int idx_a = 0; idx_a < hand_top; idx_a++)
+    {
+        for (int idx_b = idx_a + 1; idx_b <= hand_top; idx_b++)
+        {
+            if (hand[idx_a] == NULL ||
+                (hand[idx_b] != NULL && hand[idx_a]->card->rank > hand[idx_b]->card->rank))
+            {
+                swap_cards_in_hand(idx_a, idx_b);
+            }
+        }
+    }
+}
+
+static inline enum HandType hand_get_type(void)
+{
+    enum HandType res_hand_type = NONE;
+
+    // Idk if this is how Balatro does it but this is how I'm doing it
+    if (hand_selections == 0 || hand_state == HAND_DISCARD)
+    {
+        res_hand_type = NONE;
+        return res_hand_type;
+    }
+
+    res_hand_type = HIGH_CARD;
+
+    u8 suits[NUM_SUITS];
+    u8 ranks[NUM_RANKS];
+    get_hand_distribution(ranks, suits);
+
+    // Check for flush
+    if (hand_contains_flush(suits))
+        res_hand_type = FLUSH;
+
+    // Check for straight
+    if (hand_contains_straight(ranks))
+    {
+        if (res_hand_type == FLUSH)
+            res_hand_type = STRAIGHT_FLUSH;
+        else
+            res_hand_type = STRAIGHT;
+    }
+
+    // The following can be optimized better but not sure how much it matters
+    u8 n_of_a_kind = hand_contains_n_of_a_kind(ranks);
+
+    if (n_of_a_kind >= 5)
+    {
+        if (res_hand_type == FLUSH)
+        {
+            return FLUSH_FIVE;
+        }
+        return FIVE_OF_A_KIND;
+    }
+
+    // Check for royal flush vs regular straight flush
+    if (res_hand_type == STRAIGHT_FLUSH)
+    {
+        if (ranks[TEN] && ranks[JACK] && ranks[QUEEN] && ranks[KING] && ranks[ACE])
+            return ROYAL_FLUSH;
+        return STRAIGHT_FLUSH;
+    }
+
+    if (n_of_a_kind == 4)
+    {
+        return FOUR_OF_A_KIND;
+    }
+
+    if (n_of_a_kind == 3 && hand_contains_full_house(ranks))
+    {
+        return FULL_HOUSE;
+    }
+
+    // Flush and Straight are more valuable than the remaining hand types, so return them now
+    if (res_hand_type == FLUSH)
+    {
+        if (n_of_a_kind >= 5)
+        {
+            return FLUSH_HOUSE;
+        }
+        return FLUSH;
+    }
+    if (res_hand_type == STRAIGHT)
+    {
+        return STRAIGHT;
+    }
+
+    if (n_of_a_kind == 3)
+    {
+        return THREE_OF_A_KIND;
+    }
+
+    if (n_of_a_kind == 2)
+    {
+        if (hand_contains_two_pair(ranks))
+        {
+            return TWO_PAIR;
+        }
+        return PAIR;
+    }
+
+    return res_hand_type; // should be HIGH_CARD
+}
+
+static void print_hand_type(const char* hand_type_str)
+{
+    if (hand_type_str == NULL)
+        return; // NULL-checking paranoia
+    tte_printf(
+        "#{P:%d,%d; cx:0x%X000}%s",
+        HAND_TYPE_RECT.left,
+        HAND_TYPE_RECT.top,
+        TTE_WHITE_PB,
+        hand_type_str
+    );
+}
+
+static void set_hand(void)
+{
+    tte_erase_rect_wrapper(HAND_TYPE_RECT);
+    hand_type = hand_get_type();
+
+    HandValues hand = hand_base_values[hand_type];
+
+    chips = hand.chips;
+    mult = hand.mult;
+
+    print_hand_type(hand.display_name);
+    display_chips();
+    display_mult();
+}
 
 static bool sort_by_suit = false;
 
@@ -307,7 +579,7 @@ static bool card_selected_instead_of_moved = false;
 static const int card_swap_time_threshold = 6;
 static uint selection_hit_timer = TM_ZERO;
 
-bool game_playing_hand_row_on_selection_changed(
+static bool game_playing_hand_row_on_selection_changed(
     SelectionGrid* selection_grid,
     int row_idx,
     const Selection* prev_selection,
@@ -318,7 +590,7 @@ bool game_playing_hand_row_on_selection_changed(
     int next_card_idx = UNDEFINED;
 
     // Do not use FRAMES(x) here as we are counting real frames ignoring game speed
-    card_moved_too_fast = (timer - selection_hit_timer) < card_swap_time_threshold;
+    card_moved_too_fast = (get_timer() - selection_hit_timer) < card_swap_time_threshold;
 
     if (prev_selection->y == GAME_PLAYING_HAND_SEL_Y)
     {
@@ -383,11 +655,14 @@ bool game_playing_hand_row_on_selection_changed(
     return true;
 }
 
-void game_playing_hand_row_on_key_transit(SelectionGrid* selection_grid, Selection* selection)
+static void game_playing_hand_row_on_key_transit(
+    SelectionGrid* selection_grid,
+    Selection* selection
+)
 {
     if (key_hit(SELECT_CARD))
     {
-        selection_hit_timer = timer;
+        selection_hit_timer = get_timer();
     }
     else if (key_released(SELECT_CARD))
     {
@@ -416,7 +691,7 @@ static inline void game_playing_button_set_highlight(int btn_idx, bool highlight
     button_set_highlight(&game_playing_buttons[btn_idx], highlight);
 }
 
-bool game_playing_button_row_on_selection_changed(
+static bool game_playing_button_row_on_selection_changed(
     SelectionGrid* selection_grid,
     int row_idx,
     const Selection* prev_selection,
@@ -441,7 +716,7 @@ bool game_playing_button_row_on_selection_changed(
     return true;
 }
 
-void game_playing_button_row_on_key_hit(SelectionGrid* selection_grid, Selection* selection)
+static void game_playing_button_row_on_key_hit(SelectionGrid* selection_grid, Selection* selection)
 {
     if (key_hit(SELECT_CARD))
     {
@@ -474,11 +749,18 @@ static void hand_change_sort(void)
     sort_cards();
 }
 
-extern bool can_play_hand(void)
+static bool can_play_hand(void)
 {
     if (hand_state != HAND_SELECT || hand_selections == 0)
         return false;
     return true;
+}
+
+static bool can_discard_hand(void)
+{
+    int discards = get_discards();
+    enum HandState hand_state = get_hand_state();
+    return (discards > 0 && hand_state == HAND_SELECT && hand_selections > 0);
 }
 
 /**
@@ -612,7 +894,7 @@ static inline void card_in_hand_loop_handle_discard_and_shuffling(
                 // This technically isn't drawing cards, I'm just reusing the variable
                 cards_drawn++;
                 sound_played = false;
-                timer = TM_ZERO;
+                reset_timer();
 
                 *hand_y = hand[card_idx]->sprite_object->y;
                 *hand_x = hand[card_idx]->sprite_object->x;
@@ -640,7 +922,7 @@ static inline void card_in_hand_loop_handle_discard_and_shuffling(
         *hand_x = *hand_x + (int2fx(card_idx) - int2fx(hand_top) / 2) * -HAND_SPACING_LUT[hand_top];
     }
 
-    if (card_idx == 0 && discarded_card == false && timer % FRAMES(10) == 0)
+    if (card_idx == 0 && discarded_card == false && get_timer() % FRAMES(10) == 0)
     {
         // This is never reached in the case of HAND_SHUFFLING. Not sure why but that's how it's
         // supposed to be.
@@ -648,7 +930,7 @@ static inline void card_in_hand_loop_handle_discard_and_shuffling(
         sound_played = false;
         cards_drawn = 0;
         hand_selections = 0;
-        timer = TM_ZERO;
+        reset_timer();
         *break_loop = true;
         return;
     };
@@ -881,7 +1163,7 @@ static inline bool game_round_is_over(void)
 // returns true if the current card has been discarded
 static bool play_ended_played_cards_update(int played_idx)
 {
-    if (!discarded_card && timer > FRAMES(40))
+    if (!discarded_card && get_timer() > FRAMES(40))
     {
         // play the sound only once per card, when it is pushed off-screen to the right
         if (!sound_played)
@@ -922,7 +1204,7 @@ static bool play_ended_played_cards_update(int played_idx)
                 played_top = -1; // Reset the played stack
                 scored_card_index = 0;
                 _joker_scored_itr = list_itr_create(&_owned_jokers_list);
-                timer = TM_ZERO;
+                reset_timer();
             }
 
             return true; // return early to avoid accessing played[played_idx] == NULL
@@ -939,6 +1221,7 @@ static bool play_ended_played_cards_update(int played_idx)
 static inline void play_starting_played_cards_update(int played_idx)
 {
     bool card_selected = card_object_is_selected(played[played_top - scored_card_index]);
+    uint timer = get_timer();
     if (played_idx == played_top && (timer % FRAMES(10) == 0 || !card_selected) &&
         timer > FRAMES(40))
     {
@@ -947,7 +1230,7 @@ static inline void play_starting_played_cards_update(int played_idx)
         if (scored_card_index == 0)
         {
             _joker_scored_itr = list_itr_create(&_owned_jokers_list);
-            timer = TM_ZERO;
+            reset_timer();
             play_state = PLAY_BEFORE_SCORING;
         }
     }
@@ -979,6 +1262,7 @@ static inline bool play_before_scoring_cards_update(void)
 // returns true if the scoring loop has returned early
 static inline bool play_scoring_cards_update(void)
 {
+    uint timer = get_timer();
     if (timer % FRAMES(30) == 0 && timer > FRAMES(40))
     {
         // We are about to score played Cards.
@@ -1046,6 +1330,7 @@ static inline bool play_scoring_cards_update(void)
 // returns true if the scoring loop has returned early
 static inline bool play_scoring_card_jokers_update(void)
 {
+    uint timer = get_timer();
     if (timer % FRAMES(30) == 0 && timer > FRAMES(40))
     {
         tte_erase_rect_wrapper(PLAYED_CARDS_SCORES_RECT);
@@ -1091,6 +1376,7 @@ static inline bool play_scoring_card_jokers_update(void)
 // returns true if the scoring loop has returned early
 static inline bool play_scoring_held_cards_update(int played_idx)
 {
+    uint timer = get_timer();
     if (played_idx == 0 && (timer % FRAMES(30) == 0) && timer > FRAMES(40))
     {
         tte_erase_rect_wrapper(HELD_CARDS_SCORES_RECT);
@@ -1122,6 +1408,7 @@ static inline bool play_scoring_held_cards_update(int played_idx)
 // returns true if the scoring loop has returned early
 static inline bool play_scoring_independent_jokers_update(int played_idx)
 {
+    uint timer = get_timer();
     if (played_idx == 0 && (timer % FRAMES(30) == 0) && timer > FRAMES(40))
     {
 
@@ -1144,6 +1431,7 @@ static inline bool play_scoring_independent_jokers_update(int played_idx)
 // Trigger hand end effect for all jokers once they are done scoring
 static inline bool play_scoring_hand_scored_end_update(int played_idx)
 {
+    uint timer = get_timer();
     if (played_idx == 0 && (timer % FRAMES(30) == 0) && timer > FRAMES(40))
     {
 
@@ -1160,7 +1448,7 @@ static inline bool play_scoring_hand_scored_end_update(int played_idx)
             return true;
         }
 
-        timer = TM_ZERO;
+        reset_timer();
         play_state = PLAY_ENDING;
     }
 
@@ -1171,6 +1459,7 @@ static inline bool play_scoring_hand_scored_end_update(int played_idx)
 // sequentially
 static inline void play_ending_played_cards_update(int played_idx)
 {
+    uint timer = get_timer();
     bool card_selected = card_object_is_selected(played[played_top - scored_card_index]);
     if (played_idx == played_top && (timer % FRAMES(10) == 0 || !card_selected) &&
         timer > FRAMES(40))
@@ -1189,7 +1478,7 @@ static inline void play_ending_played_cards_update(int played_idx)
                 CHIPS_ACCUM_SFX_PITCH_RATIO * MM_BASE_PITCH_RATE,
                 SFX_DEFAULT_VOLUME
             );
-            timer = TM_ZERO;
+            reset_timer();
             play_state = PLAY_ENDED;
         }
     }
@@ -1334,7 +1623,7 @@ static inline void game_playing_process_input_and_state(void)
             );
         }
     }
-    else if (play_state == PLAY_ENDED && timer % FRAMES(TM_SCORE_LERP_INTERVAL) == 0)
+    else if (play_state == PLAY_ENDED && get_timer() % FRAMES(TM_SCORE_LERP_INTERVAL) == 0)
     {
         /* Using fixed point in case the score is lower than NUM_SCORE_LERP_STEPS and then
          * then the division rounds it down to 0 and it's never added to the total.
@@ -1370,7 +1659,7 @@ static inline void game_playing_process_card_draw()
 {
     if (hand_state == HAND_DRAW && cards_drawn < hand_size)
     {
-        if (timer % FRAMES(10) == 0) // Draw a card every 10 frames
+        if (get_timer() % FRAMES(10) == 0) // Draw a card every 10 frames
         {
             cards_drawn++;
             card_draw();
@@ -1380,7 +1669,7 @@ static inline void game_playing_process_card_draw()
     {
         hand_state = HAND_SELECT; // Change the hand state to select after drawing all the cards
         cards_drawn = 0;
-        timer = TM_ZERO;
+        reset_timer();
     }
 }
 
@@ -1388,7 +1677,7 @@ static inline void game_playing_discarded_cards_loop(void)
 {
     // Discarded cards loop (mainly for shuffling)
     if (hand_get_size() == 0 && hand_state == HAND_SHUFFLING && discard_top >= -1 &&
-        timer > FRAMES(10))
+        get_timer() > FRAMES(10))
     {
         // Change the background to the round end background. This is how it works in Balatro, so
         // I'm doing it this way too.
@@ -1549,6 +1838,7 @@ static inline void cards_in_hand_update_loop(void)
                     hand_x =
                         hand_x + (int2fx(i) - int2fx(hand_top) / 2) * -HAND_SPACING_LUT[hand_top];
                     hand_y += int2fx(24);
+                    uint timer = get_timer();
 
                     if (card_object_is_selected(hand[i]) && discarded_card == false &&
                         timer % FRAMES(10) == 0)
@@ -1577,7 +1867,7 @@ static inline void cards_in_hand_update_loop(void)
                         hand_state = HAND_PLAYING;
                         cards_drawn = 0;
                         hand_selections = 0;
-                        timer = TM_ZERO;
+                        reset_timer();
                         scored_card_index = played_top + 1;
 
                         select_cards_in_played_hand();
@@ -1603,6 +1893,7 @@ static inline void game_playing_ui_text_update(void)
 {
     static int last_hand_size = 0;
     static int last_deck_size = 0;
+    enum BackgroundId background = get_background();
 
     if (last_hand_size != hand_get_size() || last_deck_size != deck_get_size())
     {
@@ -1652,7 +1943,7 @@ static inline void game_playing_process_flaming_score(void)
 
     if (score_flames_active)
     {
-        if (timer % SCORE_FLAMES_ANIM_FREQ == 0)
+        if (get_timer() % SCORE_FLAMES_ANIM_FREQ == 0)
         {
             Rect frame_rect = SCORE_FLAME_FRAMES_START;
             flame_score_frame = (flame_score_frame + 1) % NUM_SCORE_FLAMES_FRAMES;
@@ -1675,7 +1966,9 @@ void game_playing_change_background(enum BackgroundId current_background)
     if (current_background != BG_CARD_SELECTING)
     {
         change_background(BG_CARD_SELECTING);
-        background = BG_CARD_PLAYING;
+        set_background(BG_CARD_PLAYING
+        ); // Not sure why this is set? I'm pretty sure at the end of the change_background function
+           // it sets it to BG_CARD_SELECTING anyway. -emiyl
     }
 
     REG_WIN0V = (REG_WIN0V << 8) | 0xA0; // Set window 0 bottom to 160
